FRONT-6 updated after review (added pointed cursor, fixed the top margin, replaced the text where suggested - still waiting for Jacob's confirmation regarding the placement of the mapped functions from Header
