+++ src/components/Footer/Footer.css
@@ -2,7 +2,7 @@
     position: fixed;
     bottom: 0;
     width: 100%;
-    height: 70px;
+    height: 85px;
     background-color: azure;
     display: grid;
     grid-template-columns: repeat(3, 1fr);
@@ -11,16 +11,25 @@
     grid-row-gap: 0px;
 }
 
+.git-links p {
+    margin-right: 25px;
+}
+
 a {
     padding-right: 5px;
 }
+
+a img {
+    width: 30px;
+    padding-bottom: 5px;
+}
     
 .copyright { grid-area: 1 / 1 / 2 / 2; margin-top: 50px;}
-.disclaimer { grid-area: 1 / 2 / 2 / 3; text-align: center; margin-top: 30px;}
+.disclaimer { grid-area: 1 / 2 / 2 / 3; text-align: center;}
 .git-links { grid-area: 1 / 3 / 2 / 4; text-align: right; padding-right: 20px;}
 
 @media (max-width: 768px) {
-    .disclaimer {
+    .disclaimer-desktop {
         display: none;
     }
 }
