+++ src/components/Header/Burger.jsx
@@ -24,6 +24,7 @@ const StyledBurger = styled.div`
         border-radius: 10px;
         transform-origin: 1px;
         transition: all 0.3s linear;
+        cursor: pointer;
 
         &:nth-child(1) {
             transform: ${({ open }) => open ? 'rotate(45deg)' : 'rotate(0)'};
