+++ src/components/Header/Header.css
@@ -1,14 +1,17 @@
 .header {
     position: sticky;
     top:0;
-    padding:40px;
   }
 
 .logo {
-  padding: 30px 0;
+  /* padding: 30px 0; */
 }
 
-.Nav {
+.logo img {
+  width: 80px;
+}
+
+.nav {
   width: 100%;
   height: 65px;
   border-bottom: 2px solid #f1f1f1;
